update
git-svn-id: http://svn.tudos.org/repos/tudos/trunk@450 a704ac0b-3a55-4d43-a2a9-7be6f07c34fb

+++ l4/pkg/dde/linux26/include/ARCH-x86/asm-i386/io.h
@@ -88,7 +88,7 @@
  
 static inline unsigned long virt_to_phys(volatile void * address)
 {
-	return __pa(address);
+	return __pa((unsigned long)address);
 }
 
 /**
